requirements.txt
transformers
# torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torch
elasticsearch
pyyaml
tqdm